+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this library are documented in this file.
 
 ### API Changes
 
+- Add `pyiem.reference.StationAttributes` enum to hold database attribute
+  keys used.
 - Discontinue raw SPS product text within the database.
 - Ensure raw text products into `WMOProduct` end with a line feed.
 - Stop postprocessing the `MapPlot` figure into 8bit colorspace.